@@ -12,15 +12,15 @@
         }
     ],
     "require": {
-        "php": ">=7.0.0",
-        "symfony/console": "^3.2",
-        "symfony/filesystem": "^3.2",
+        "php": ">=8.0",
+        "symfony/console": "^6.1",
+        "symfony/filesystem": "^6.1",
         "herrera-io/phar-update": "2.*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7",
         "phing/phing": "2.*",
-        "mikey179/vfsStream": "^1.6"
+        "phpunit/phpunit": "^9.5",
+        "mikey179/vfsstream": "^1.6"
     },
     "autoload": {
         "psr-4": {
