@@ -9,11 +9,12 @@
 
 namespace TeamNeusta\Hosts\Tests\Services;
 
+use PHPUnit\Framework\TestCase;
 use TeamNeusta\Hosts\Services\InitService;
 use TeamNeusta\Hosts\Services\Provider\Filesystem;
 use org\bovigo\vfs\vfsStream;
 
-class InitServiceTest extends \PHPUnit_Framework_TestCase
+class InitServiceTest extends TestCase
 {
     /**
      * @var \TeamNeusta\Hosts\Services\Provider\Filesystem | \PHPUnit_Framework_MockObject_MockObject
@@ -23,7 +24,7 @@ class InitServiceTest extends \PHPUnit_Framework_TestCase
     /**
      * Public setUp.
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
